# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AudioMetadata": {
        "properties": {
          "channels": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "sample_rate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Rate"
          }
        },
        "title": "AudioMetadata",
        "type": "object"
      },
...
      "CreateTaggingRequest": {
        "properties": {
          "audio_key": {
            "description": "Key of an ingested clip under audio/",
            "title": "Audio Key",
            "type": "string"
          },
          "model": {
            "default": "cnn14-32k",
            "enum": [
              "cnn14-32k",
              "cnn14-16k"
            ],
            "title": "Model",
            "type": "string"
          },
          "top_k": {
            "default": 10,
            "enum": [
              5,
              10,
              15,
              20
            ],
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "audio_key"
        ],
        "title": "CreateTaggingRequest",
        "type": "object"
      },
...
      "EditTaggingRequest": {
        "description": "Edit an existing Tagging. `audio_key` is the read-only identity.",
        "properties": {
          "audio_key": {
            "title": "Audio Key",
            "type": "string"
          },
          "model": {
            "default": "cnn14-32k",
            "enum": [
              "cnn14-32k",
              "cnn14-16k"
            ],
            "title": "Model",
            "type": "string"
          },
          "top_k": {
            "default": 10,
            "enum": [
              5,
              10,
              15,
              20
            ],
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "audio_key"
        ],
        "title": "EditTaggingRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Label": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "probability": {
            "title": "Probability",
            "type": "number"
          }
        },
        "required": [
          "label",
          "probability"
        ],
        "title": "Label",
        "type": "object"
      },
...
      "RunTaggingRequest": {
        "description": "Re-tag with the Tagging's stored parameters (no form).",
        "properties": {
          "audio_key": {
            "title": "Audio Key",
            "type": "string"
          }
        },
        "required": [
          "audio_key"
        ],
        "title": "RunTaggingRequest",
        "type": "object"
      },
      "Tagging": {
        "description": "Full Tagging detail, including the embedding and its summary stats.",
        "properties": {
          "audio_key": {
            "title": "Audio Key",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "embedding": {
            "items": {
              "type": "number"
            },
            "title": "Embedding",
            "type": "array"
          },
          "embedding_dim": {
            "title": "Embedding Dim",
            "type": "integer"
          },
          "embedding_l2_norm": {
            "title": "Embedding L2 Norm",
            "type": "number"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/Label"
            },
            "title": "Labels",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "source_metadata": {
            "$ref": "#/components/schemas/AudioMetadata"
          },
          "tag_key": {
            "title": "Tag Key",
            "type": "string"
          },
          "tagged_at": {
            "format": "date-time",
            "title": "Tagged At",
            "type": "string"
          },
          "top_k": {
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "audio_key",
          "tag_key",
          "model",
          "top_k",
          "labels",
          "tagged_at",
          "embedding",
          "embedding_dim",
          "embedding_l2_norm",
          "source_metadata"
        ],
        "title": "Tagging",
        "type": "object"
      },
      "TaggingSummary": {
        "description": "List-view projection of a Tagging (no embedding vector).",
        "properties": {
          "audio_key": {
            "title": "Audio Key",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/Label"
            },
            "title": "Labels",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "tag_key": {
            "title": "Tag Key",
            "type": "string"
          },
          "tagged_at": {
            "format": "date-time",
            "title": "Tagged At",
            "type": "string"
          },
          "top_k": {
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "audio_key",
          "tag_key",
          "model",
          "top_k",
          "labels",
          "tagged_at"
        ],
        "title": "TaggingSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local API for PANNs AudioSet Tagging: ingest audio clips to Backblaze B2, tag them locally with PANNs (527-class AudioSet labels + a 2048-dim embedding), and store every artifact \u2014 audio, per-clip tag JSON, and the labels_index.jsonl manifest \u2014 in B2 over the S3-compatible API. This contract documents the local API, not a hosted public endpoint.",
    "title": "PANNs AudioSet Tagging API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/taggings": {
      "delete": {
        "operationId": "delete_tagging_endpoint_taggings_delete",
        "parameters": [
          {
            "in": "query",
            "name": "audio_key",
            "required": true,
            "schema": {
              "title": "Audio Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Tagging Endpoint",
        "tags": [
          "tagging"
        ]
      },
      "get": {
        "operationId": "list_taggings_endpoint_taggings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaggingSummary"
                  },
                  "title": "Response List Taggings Endpoint Taggings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Taggings Endpoint",
        "tags": [
          "tagging"
        ]
      },
      "post": {
        "operationId": "create_tagging_endpoint_taggings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaggingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tagging"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tagging Endpoint",
        "tags": [
          "tagging"
        ]
      }
    },
    "/taggings/detail": {
      "get": {
        "operationId": "get_tagging_endpoint_taggings_detail_get",
        "parameters": [
          {
            "in": "query",
            "name": "audio_key",
            "required": true,
            "schema": {
              "title": "Audio Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tagging"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tagging Endpoint",
        "tags": [
          "tagging"
        ]
      }
    },
    "/taggings/edit": {
      "post": {
        "operationId": "edit_tagging_endpoint_taggings_edit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditTaggingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tagging"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit Tagging Endpoint",
        "tags": [
          "tagging"
        ]
      }
    },
...
    "/taggings/run": {
      "post": {
        "operationId": "run_tagging_endpoint_taggings_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunTaggingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tagging"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Tagging Endpoint",
        "tags": [
          "tagging"
        ]
      }
    },
...
  }
}